requirements.txt
django-rest-framework
pillow
django-crispy-forms
crispy-bootstrap4
python-dotenv

django-allauth
django-allauth[socialaccount]
stripe